incremental commit toward a full AMP tracker
added HeunEuler predictor method
added a field to the NumTraits, which is NumFuzzyDigits
added functions for computing the right hand sides of the AMP criteria, B and C
added overload of Euler which returns the norm of Jacobian, and its inverse
added Predict() overloads, which return the norms of J and J inverse, as well as the size_proportion, $a$ from AMP2 paper.
added a function for querying whether a predictor can return an error estimate.
continued refinement of the Tracker class.
some massaging of the config options for tracking.