now with MultiObserver, and the ability to glue together several observers into a bigger one!  yay!
the MultiObserver is implemented using a std::tuple of observers, of types fed into the template instantiation.  It uses Boost.Fusion to iterate over the members of the tuple, and a lambda applying the member function to each.  A basic test using the GoryDetailLogger and a FailedStep observer are provided in this commit, though the examples are likely to change.