changed signature of internal FreshEval function, to resolve performance issue.
performance issue was as follows:

the fresh eval function was overloaded using a dummy variable passed down through the FreshEval.  this unused variable was initialized at each and every call to the top level Eval, and copied down the chain.  i opted to remove this dummy variable, and replace it by a bertini1 style _d or _mp suffix, which is looked up by a detail selector.

the signature has been further enhanced to pass by reference, rather than value.  this should help solve the performance issues caused by frequent creation and destruction of shared pointers to nodes when evaluating and differentiating.