fixes to the type system, and precision control.
tests were failing for fixed precision endgames, due to a circular call to EnsureAtUniformPrecision.  This same bug bit the adaptive endgames, and was fixed a few commits ago.

also, a few improvements to naming of variables in the System::TimeDerivative function.
also, testing improvements for the endgames.  now test multiple ambient precisions for the EG test suites.