added permanent temporaries to the sum and multiply operators.
the new temporaries have their precision change along with the other internal data.  this also permits re-use on each evaluation call, eliminating the generation of a temporary.

also a few cosmetic changes to the FreshEvalInPlace calls for arithmetic operators.

there is a bug somewhere, causing problems in the execution of
.libs/endgames_test --run_test=amp_powerseries_endgame_generic_tests_ambient_precision_16/griewank_osborne