now find lib in same place as include for bertini.
i still think this file could be a lot better, because it makes some assumptions about the folder layout.  namely, it assumes it goes /path/to/include and /path/to/lib, and that /path/to is the same for both.  if you can improve this, please do. otherwise, if this ever causes an issue, i apologize, and still ask for your help.