corrected the `Eigen::NumTraits` for `expression`s of `mpfr_float`s.
This is for compatibility with `et_on` for Boost.Multiprecision's number type.  The function from Eigen giving problems was `abs2`, which was returning `x*x` into an expression of more complicated type.  Basically, if the template provided in this commit is not present, then when Eigen looks up the `Real` type for `expression`s of `mpfr_float`s, the Real type is the expression's type, which cannot be instantiated.  So compilation failed.  This template provides the lookup that the `Real` type for expressions of `mpfr_float`s is `mpfr_float`.