replaced MakeVariable, MakeInteger, MakeRational, etc.
having the new factory wrappers, intended to be modified later to allow a formal manager or factory to be implemented, I replaced the calls to std::make_shared<node::Variable>() with MakeVariable.  Similarly for MakeInteger.

I think a MakeFunction could be good.  this level of abstraction and formalization will be useful in extracting performance gains from the system, i think.