corrected Visit for SumOperator
we had been using the `signs` incorrectly -- there's actually one sign per operand, indicating what the sign in the sum is.  so, if the first sign is negative (false), then we start by negating.

it's very likely the multiplication/division operator's visit is incorrect, cuz I think I was making the same mistake.