added explicit setting of eval method
these tests depend on eval method to be FunctionTree.  this is a side-effect of setting variable values NOT propagating to the Node variables when eval method is SLP. Like, you're using an SLP.

down to 8 failures, but I want to re-write the differentiate and differentiate_wrt_var test suites to be much more explicit and thorough.  right now the suites use Systems to wrap around, and the systems and parsed from strings using classic parser.  but this is testing more than just differentiation.  it needs to be separated.