commented out lines that just weren't working for me
they cause install errors because the files don't go to the right places.  leave it to pip.

this was a remnant of an effort to get autotools to correctly make a distfile, so we could use it with conda.  fail.  we're working on replacing autotools with cmake, but this is work in progress.

sorry for the noise.